Added questions and comments
Added comments and (mostly) questions on the model code. Main problems were that I couldn't understand the parameters of most functions and classes, and the Encoder doesn't follow the architecture exactly in the paper. Also, he whereabouts of the style encoder (we suspect it's using the WaveNet Vocoder specified here https://github.com/r9y9/wavenet_vocoder, but couldn't really find it) are yet unclear.